+++ packages/utils-tests-helpers/src/hint-runner.ts
@@ -80,6 +80,55 @@ const findPosition = (source: string, position: MatchProblemLocation): ProblemLo
 };
 
 /**
+ * Compares two location to see if they have the same start/end values
+ */
+export const comparePositions = (position1: ProblemLocation, position2: ProblemLocation): boolean => {
+    if (position1.line !== position2.line) {
+        return false;
+    }
+    if (position1.column !== position2.column) {
+        return false;
+    }
+    if (position1.endLine !== position2.endLine) {
+        return false;
+    }
+    if (position1.endColumn !== position2.endColumn) {
+        return false;
+    }
+
+    return true;
+};
+
+/**
+ * Translates line and column start and end values to offset values.
+ */
+export const positionToOffset = (position: ProblemLocation, document: string): number[] => {
+    const {column, endColumn, endLine, line} = position;
+
+    if (typeof endLine !== 'number' || typeof endColumn !== 'number') {
+        return [-1, -1];
+    }
+    let startOffset = column;
+    let endOffset = endColumn;
+
+    const regex = /(\r\n|\r|\n)/gm;
+
+    for (let i = 0; i < endLine && regex.exec(document); i++) {
+        const curLineOffset = regex.lastIndex || -1;
+
+        if (i === line - 1) {
+            startOffset = curLineOffset + column;
+        }
+
+        if (i === endLine - 1) {
+            endOffset = curLineOffset + endColumn;
+        }
+    }
+
+    return [startOffset, endOffset];
+};
+
+/**
  * Get the source code for the provided resource.
  * Returns the empty string if resource was invalid.
  */
@@ -192,7 +241,7 @@ const validateResults = (t: ExecutionContext<HintRunnerContext>, sources: Map<st
          * information to the report that matches the closest.
          */
 
-        const { documentation, location, message, resource, severity } = result;
+        const { documentation, fixes, location, message, resource, severity } = result;
 
         const filteredByMessage = reportsCopy.filter((report) => {
             if (typeof report.message === 'string') {
@@ -211,7 +260,7 @@ const validateResults = (t: ExecutionContext<HintRunnerContext>, sources: Map<st
         const filteredByDocumentationCount = filteredByMessage.filter((report) => {
             /*
              * If the report from the test doesn't ask for documentation,
-             * we don't need to macth it.
+             * we don't need to match it.
              */
             if (!report.documentation) {
                 return true;
@@ -330,7 +379,93 @@ const validateResults = (t: ExecutionContext<HintRunnerContext>, sources: Map<st
             return;
         }
 
-        const filteredBySeverity = filteredByPosition.filter((report) => {
+        const filteredByFixes = filteredByPosition.filter((report) => {
+            /*
+             * If the report from the test doesn't ask for fixes,
+             * we don't need to match it.
+             */
+            if (!report.fixes) {
+                return true;
+            }
+
+            /*
+             * If the report from the test does ask for fixes
+             * but the result doesn't provide it, then it isn't a match.
+             */
+            if (!fixes) {
+                return false;
+            }
+
+            if (Array.isArray(report.fixes)) {
+                if (report.fixes.length !== fixes.length) {
+                    return false;
+                }
+
+                for (let i = 0; i < fixes.length; i++) {
+                    const curLocation = fixes[i].location;
+                    const curText = fixes[i].text;
+                    const targetLocation = report.fixes[i].location;
+                    const targetText = report.fixes[i].text;
+
+                    if (curText !== targetText || !comparePositions(curLocation, targetLocation)) {
+                        return false;
+                    }
+                }
+
+                return true;
+            } else if (report.fixes.match) {
+                let sourceCopy = sources.get(resource);
+
+                if (!sourceCopy) {
+                    return false;
+                }
+
+                /**
+                 *  We want to sort fixes starting from the end of the document moving upwards to avoid one fix changing the offset of a subsequent fix.
+                 */
+                fixes.sort((a, b) => {
+                    const lineDiff = b.location.line - a.location.line;
+
+                    if (lineDiff === 0) {
+                        return b.location.column - a.location.column;
+                    }
+
+                    return lineDiff;
+                });
+
+                for (const fix of fixes) {
+                    let startOffset = fix.location.startOffset;
+                    let endOffset = fix.location.endOffset;
+
+
+                    if (!startOffset || !endOffset) {
+                        const document = sources.get(resource);
+
+                        if (document) {
+                            [startOffset, endOffset] = positionToOffset(fix.location, document);
+                            if (startOffset < 0 || endOffset < 0) {
+                                return false;
+                            }
+                        } else {
+                            return false;
+                        }
+                    }
+
+                    sourceCopy = sourceCopy.substring(0, startOffset) + fix.text + sourceCopy.substring(endOffset);
+                }
+                t.is(sourceCopy, report.fixes.match);
+
+                return sourceCopy === report.fixes.match;
+            }
+
+            return false;
+        });
+
+        if (filteredByFixes.length === 0) {
+            t.fail(`The fix ${JSON.stringify(fixes)} does not match any report for "${message}"`);
+        }
+
+        const filteredBySeverity = filteredByFixes.filter((report) => {
             // Not all reports in the test have a severity
             if (typeof report.severity === 'undefined') {
                 return true;