+++ packages/utils-tests-helpers/src/hint-test-type.ts
@@ -1,6 +1,6 @@
 import { ExecutionContext } from 'ava';
 
-import { ProblemLocation, Severity, ProblemDocumentation } from '@hint/utils-types';
+import { CodeFix, ProblemLocation, Severity, ProblemDocumentation } from '@hint/utils-types';
 
 export type MatchProblemLocation = {
     /** A substring matching the location of the problem. */
@@ -14,6 +14,7 @@ export type Report = {
     position?: ProblemLocation | MatchProblemLocation;
     severity?: Severity;
     documentation?: ProblemDocumentation[];
+    fixes?: CodeFix[] | { match: string };
 };
 
 export type HintTest = {